Move icons to parent `components` directory
