+++ assets/src/block-validation/components/amp-document-status/index.js
@@ -11,7 +11,7 @@ import { useDispatch, useSelect } from '@wordpress/data';
 import AMPValidationErrorsKeptIcon from '../../../../images/amp-validation-errors-kept.svg';
 import BellIcon from '../../../../images/bell-icon.svg';
 import { BLOCK_VALIDATION_STORE_KEY } from '../../store';
-import { StatusIcon } from '../icon';
+import { StatusIcon } from '../../../components/icon';
 import { SidebarNotification } from '../sidebar-notification';
 import { useAMPDocumentToggle } from '../../hooks/use-amp-document-toggle';
 import { useErrorsFetchingStateChanges } from '../../hooks/use-errors-fetching-state-changes';
