+++ assets/src/components/icon/test/icon.js
